update the rotation function so the so u dont have to keep pushing the buttons in order to rotate

+++ DataTypes.hs
@@ -5,7 +5,7 @@ data Player = Player
     { degree :: Float            --  the degree will be W.R.T X-axis like this   (>)  <- space ship at degree 0
     , plSpeed  :: (Float, Float)  -- speed W.R.T (x and y axes)
     , plLocation :: (Float, Float) -- location W.R.T (x and y axes)
-    , rotatingBy :: Int
+    , rotatingBy :: Float
     , firingSpeed :: Int               -- fire speed for more fun :D
     , isrotating :: Bool
     , isFiring :: Bool
